fs: fs_seek may return new position on success
The commit adds FS_SEEK_RET_POS flag that, in combination with 'whence'
identifier, makes fs_seek return on success a new position within file
instead of 0.

+++ include/fs/fs_sys.h
@@ -46,7 +46,7 @@ struct fs_file_system_t {
 	ssize_t (*read)(struct fs_file_t *filp, void *dest, size_t nbytes);
 	ssize_t (*write)(struct fs_file_t *filp,
 					const void *src, size_t nbytes);
-	int (*lseek)(struct fs_file_t *filp, off_t off, int whence);
+	off_t (*lseek)(struct fs_file_t *filp, off_t off, int whence);
 	off_t (*tell)(struct fs_file_t *filp);
 	int (*truncate)(struct fs_file_t *filp, off_t length);
 	int (*sync)(struct fs_file_t *filp);
